requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
bcrypt
pandas
openpyxl
python-docx
comtypes
pywin32
PyPDF2